requirements.txt
streamlit>=1.32.0
requests>=2.31.0
pandas>=2.2.0
plotly>=5.18.0
numpy>=1.26.4
python-dotenv>=1.0.0
nba_api>=1.4.0
